+++ RELEASES.md
@@ -152,6 +152,7 @@ Upcoming
 - Bugfix Base64 encoding/decoding not working in JdbcQuerySender when streaming to servlet
 - Add interactive favicon for new GUI
 - Larva: show diffMessage (hidden) for "scenario passed/failed"
+- Larva: add httpRequest parameter
 
 
 ### Non backwards compatible changes